+++ README.md
@@ -17,24 +17,24 @@ A fast port scanner written in python with a focus on reliability and simplicity
 
 |     | Syntax                                                      | Examples                                     |
 |:----|:------------------------------------------------------------|:---------------------------------------------|
-| 1.  | `python3 scanner.py <ip>`                                   | `python3 scanner.py 192.168.0.1`             |
-| 2.  | `python3 scanner.py <ip> <verbose>`                         | `python3 scanner.py 192.168.0.1 -v`          |
-| 3.  | `python3 scanner.py <ip> <port_start> <port_end>`           | `python3 scanner.py 192.168.0.1 150 1333`    |
-| 4.  | `python3 scanner.py <ip> <port_start> <port_end> <verbose>` | `python3 scanner.py 192.168.0.1 150 1333 -v` |
+| 1.  | `python3 scanner.py --host <ip>`                                   | `python3 scanner.py --host 192.168.0.1`             |
+| 2.  | `python3 scanner.py --host <ip> <verbose>`                         | `python3 scanner.py --host 192.168.0.1 -v`          |
+| 3.  | `python3 scanner.py --host <ip> -p <port_start>-<port_end>`           | `python3 scanner.py --host 192.168.0.1 -p 150-1333`    |
+| 4.  | `python3 scanner.py --host <ip> -p <port_start>-<port_end> <verbose>` | `python3 scanner.py --host 192.168.0.1 -p 150-1333 -v` |
 
 #### Port specific:
 
-- Syntax: `python3 scanner.py <ip> <port_start> <port_end>`
+- Syntax: `python3 scanner.py --host <ip> -p <port_start>-<port_end>`
 
-- Example: `python3 scanner.py 192.168.1.1 0 200`
-  ![port_specific](https://user-images.githubusercontent.com/44284877/179356857-4676e09e-48ac-4cb8-96e3-2fa910a15e9a.gif)
+- Example: `python3 scanner.py --host 192.168.1.1 -p 0-200`
+  <!-- ![port_specific](https://user-images.githubusercontent.com/44284877/179356857-4676e09e-48ac-4cb8-96e3-2fa910a15e9a.gif) -->
 
 #### Verbose mode:
 
-- Syntax: `python3 scanner.py <ip> <port_start> <port_end> <verbose> `
+- Syntax: `python3 scanner.py --host <ip> -p <port_start>-<port_end> <verbose> `
 
-- Example: `python3 scanner.py 192.168.1.1 0 200 -v`
-  ![Verbose mode](https://user-images.githubusercontent.com/44284877/179357933-76ef587a-9f74-4ab7-b466-164ca4fce445.gif)
+- Example: `python3 scanner.py --host 192.168.1.1 -p 0-200 -v`
+  <!-- ![Verbose mode](https://user-images.githubusercontent.com/44284877/179357933-76ef587a-9f74-4ab7-b466-164ca4fce445.gif) -->
 
 ---
